+++ tests/Kernels/Distance/LevenshteinTest.php
@@ -52,7 +52,7 @@ class LevenshteinTest extends TestCase
     }
 
     /**
-     * @return \Generator<array>
+     * @return \Generator<array<mixed>>
      */
     public function computeProvider() : Generator
     {